+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/rmd/display/SetupChunkOptionsPopupPanel.java
@@ -47,9 +47,9 @@ public class SetupChunkOptionsPopupPanel extends ChunkOptionsPopupPanel
       server_ = server;
    }
    
-   public SetupChunkOptionsPopupPanel()
+   public SetupChunkOptionsPopupPanel(boolean isVisualEditor)
    {
-      super(false);
+      super(false, isVisualEditor);
       RStudioGinjector.INSTANCE.injectMembers(this);
       
       figureDimensionsPanel_.setVisible(false);
@@ -72,6 +72,10 @@ public class SetupChunkOptionsPopupPanel extends ChunkOptionsPopupPanel
       int row = position_.getRow();
       
       int max = display_.getRowCount();
+      
+      if (isVisualEditor_)
+         return max; // Visual Editor chunk has no closing "```" line and fills the Ace instance
+      
       for (int i = row; i < max; i++)
       {
          String line = display_.getLine(i);
@@ -164,6 +168,8 @@ public class SetupChunkOptionsPopupPanel extends ChunkOptionsPopupPanel
          
          
          token = iterator.stepForward();
+         if (token == null)
+            break; // avoid exception if we reach the buffer's end
          Position endPos = iterator.getCurrentTokenPosition();
          
          return Range.fromPoints(startPos, endPos);
@@ -187,6 +193,13 @@ public class SetupChunkOptionsPopupPanel extends ChunkOptionsPopupPanel
       }
       
       return display_.getSelectionRange();
+   }
+   
+   private void deleteLastLine()
+   {
+      display_.clearSelection();
+      display_.setCursorPosition(Position.create(display_.getRowCount() - 1, 0));
+      display_.removeLine();
    }
    
    private String joinOptions(Map<String, String> options)
@@ -230,6 +243,14 @@ public class SetupChunkOptionsPopupPanel extends ChunkOptionsPopupPanel
    @Override
    protected void synchronize()
    {
+      if (isVisualEditor_)
+      {
+         // Visual Editor chunks don't have the ending "```" line, violating assumptions
+         // in the following code. Insert a line at the end of the chunk to work around
+         // this, and remove it when we're done.
+         display_.insertCode(display_.getEnd().getEnd(), "\n```");
+      }
+      
       syncSelection();
       Map<String, String> options = new LinkedHashMap<>();
       
@@ -250,6 +271,8 @@ public class SetupChunkOptionsPopupPanel extends ChunkOptionsPopupPanel
       if (options.isEmpty())
       {
          display_.insertCode("");
+         if (isVisualEditor_)
+            deleteLastLine();
          return;
       }
       
@@ -259,6 +282,8 @@ public class SetupChunkOptionsPopupPanel extends ChunkOptionsPopupPanel
          String joined = StringUtil.collapse(options, " = ", ", ");
          String code = "knitr::opts_chunk$set(" + joined + ")\n";
          display_.insertCode(code);
+         if (isVisualEditor_)
+            deleteLastLine();
          return;
       }
       
@@ -270,6 +295,8 @@ public class SetupChunkOptionsPopupPanel extends ChunkOptionsPopupPanel
             "\n)\n";
       
       display_.insertCode(code);
+      if (isVisualEditor_)
+         deleteLastLine();
    }
    
    @Override